+++ crates/astria-bridge-withdrawer/build.rs
@@ -12,21 +12,21 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         "./astria-bridge-contracts/out/IAstriaWithdrawer.sol/IAstriaWithdrawer.json",
     )?
     .generate()?
-    .write_to_file("./src/withdrawer/ethereum/generated/astria_withdrawer_interface.rs")?;
+    .write_to_file("./src/bridge_withdrawer/ethereum/generated/astria_withdrawer_interface.rs")?;
 
     Abigen::new(
         "AstriaWithdrawer",
         "./astria-bridge-contracts/out/AstriaWithdrawer.sol/AstriaWithdrawer.json",
     )?
     .generate()?
-    .write_to_file("./src/withdrawer/ethereum/generated/astria_withdrawer.rs")?;
+    .write_to_file("./src/bridge_withdrawer/ethereum/generated/astria_withdrawer.rs")?;
 
     Abigen::new(
         "AstriaBridgeableERC20",
         "./astria-bridge-contracts/out/AstriaBridgeableERC20.sol/AstriaBridgeableERC20.json",
     )?
     .generate()?
-    .write_to_file("./src/withdrawer/ethereum/generated/astria_bridgeable_erc20.rs")?;
+    .write_to_file("./src/bridge_withdrawer/ethereum/generated/astria_bridgeable_erc20.rs")?;
 
     Ok(())
 }
