+++ crates/astria-bridge-withdrawer/src/bridge_withdrawer/ethereum/convert.rs
@@ -31,7 +31,7 @@ use serde::{
     Serialize,
 };
 
-use crate::withdrawer::ethereum::astria_withdrawer_interface::{
+use crate::bridge_withdrawer::ethereum::astria_withdrawer_interface::{
     Ics20WithdrawalFilter,
     SequencerWithdrawalFilter,
 };
@@ -191,7 +191,7 @@ mod tests {
     use asset::default_native_asset;
 
     use super::*;
-    use crate::withdrawer::ethereum::astria_withdrawer_interface::SequencerWithdrawalFilter;
+    use crate::bridge_withdrawer::ethereum::astria_withdrawer_interface::SequencerWithdrawalFilter;
 
     #[test]
     fn event_to_bridge_unlock() {
