+++ crates/astria-bridge-withdrawer/src/lib.rs
@@ -1,11 +1,11 @@
 pub(crate) mod api;
+pub mod bridge_withdrawer;
 mod build_info;
 pub(crate) mod config;
 pub mod metrics_init;
-pub mod withdrawer;
 
+#[cfg(test)]
+pub(crate) use bridge_withdrawer::astria_address;
+pub use bridge_withdrawer::BridgeWithdrawer;
 pub use build_info::BUILD_INFO;
 pub use config::Config;
-#[cfg(test)]
-pub(crate) use withdrawer::astria_address;
-pub use withdrawer::BridgeWithdrawer;
